+++ src/Abp.EntityFrameworkCore/EntityFrameworkCore/AbpDbContext.cs
@@ -76,7 +76,7 @@ namespace Abp.EntityFrameworkCore
         public virtual bool SuppressAutoSetTenantId { get; set; }
         public virtual bool SuppressAutoSetBranchId { get; set; }
 
-        protected virtual int? CurrentTenantId => GetCurrentTenantIdOrNull();
+        protected virtual long? CurrentTenantId => GetCurrentTenantIdOrNull();
         protected virtual long? CurrentBranchId => GetCurrentBranchIdOrNull();
 
         protected virtual bool IsSoftDeleteFilterEnabled => CurrentUnitOfWorkProvider?.Current?.IsFilterEnabled(AbpDataFilters.SoftDelete) == true;
@@ -589,7 +589,7 @@ namespace Abp.EntityFrameworkCore
             return null;
         }
 
-        protected virtual int? GetCurrentTenantIdOrNull()
+        protected virtual long? GetCurrentTenantIdOrNull()
         {
             if (CurrentUnitOfWorkProvider != null &&
                 CurrentUnitOfWorkProvider.Current != null)
