+++ src/Abp.EntityFrameworkCore/EntityFrameworkCore/AbpEntityFrameworkCoreExtension.cs
@@ -81,7 +81,7 @@ namespace Abp.EntityFrameworkCore
             return expression;
         }
 
-        private static int? GetCurrentTenantIdOrNull(IIocResolver iocResolver)
+        private static long? GetCurrentTenantIdOrNull(IIocResolver iocResolver)
         {
             using (var scope = iocResolver.CreateScope())
             {
