+++ src/Abp.Web.Common/Web/Models/AbpUserConfiguration/AbpUserSessionConfigDto.cs
@@ -6,11 +6,11 @@ namespace Abp.Web.Models.AbpUserConfiguration
     {
         public long? UserId { get; set; }
 
-        public int? TenantId { get; set; }
+        public long? TenantId { get; set; }
 
         public long? ImpersonatorUserId { get; set; }
 
-        public int? ImpersonatorTenantId { get; set; }
+        public long? ImpersonatorTenantId { get; set; }
 
         public MultiTenancySides MultiTenancySide { get; set; }
     }
