+++ src/Abp.Zero.Common/Application/Features/IAbpZeroFeatureValueStore.cs
@@ -4,8 +4,8 @@ namespace Abp.Application.Features
 {
     public interface IAbpZeroFeatureValueStore : IFeatureValueStore
     {
-        Task<string> GetValueOrNullAsync(int tenantId, string featureName);
-        string GetValueOrNull(int tenantId, string featureName);
+        Task<string> GetValueOrNullAsync(long tenantId, string featureName);
+        string GetValueOrNull(long tenantId, string featureName);
         Task<string> GetEditionValueOrNullAsync(int editionId, string featureName);
         string GetEditionValueOrNull(int editionId, string featureName);
         Task SetEditionFeatureValueAsync(int editionId, string featureName, string value);
