+++ src/Abp.Zero.Common/Authorization/Roles/AbpRoleBase.cs
@@ -28,7 +28,7 @@ namespace Abp.Authorization.Roles
         /// Tenant's Id, if this role is a tenant-level role. Null, if not.
         /// </summary>
         [Column("tenant_id")]
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// Unique name of this role.
@@ -67,14 +67,14 @@ namespace Abp.Authorization.Roles
             Name = Guid.NewGuid().ToString("N");
         }
 
-        protected AbpRoleBase(int? tenantId, string displayName)
+        protected AbpRoleBase(long? tenantId, string displayName)
             : this()
         {
             TenantId = tenantId;
             DisplayName = displayName;
         }
 
-        protected AbpRoleBase(int? tenantId, string name, string displayName)
+        protected AbpRoleBase(long? tenantId, string name, string displayName)
             : this(tenantId, displayName)
         {
             Name = name;
