+++ src/Abp.Zero.Common/Configuration/Setting.cs
@@ -21,7 +21,7 @@ namespace Abp.Configuration
         /// TenantId is null if this setting is not Tenant level.
         /// </summary>
         [Column("tenant_id")]
-        public virtual int? TenantId { get; set; }
+        public virtual long? TenantId { get; set; }
 
         /// <summary>
         /// UserId for this setting.
@@ -56,7 +56,7 @@ namespace Abp.Configuration
         /// <param name="userId">UserId for this setting</param>
         /// <param name="name">Unique name of the setting</param>
         /// <param name="value">Value of the setting</param>
-        public Setting(int? tenantId, long? userId, string name, string value)
+        public Setting(long? tenantId, long? userId, string name, string value)
         {
             TenantId = tenantId;
             UserId = userId;
