+++ src/Abp.Zero.Common/Localization/IApplicationLanguageManager.cs
@@ -12,25 +12,25 @@ namespace Abp.Localization
         /// Gets list of all languages available to given tenant (or null for host)
         /// </summary>
         /// <param name="tenantId">TenantId or null for host</param>
-        Task<IReadOnlyList<ApplicationLanguage>> GetLanguagesAsync(int? tenantId);
+        Task<IReadOnlyList<ApplicationLanguage>> GetLanguagesAsync(long? tenantId);
 
         /// <summary>
         /// Gets list of all active languages available to given tenant (or null for host)
         /// </summary>
         /// <param name="tenantId">TenantId or null for host</param>
-        Task<IReadOnlyList<ApplicationLanguage>> GetActiveLanguagesAsync(int? tenantId);
+        Task<IReadOnlyList<ApplicationLanguage>> GetActiveLanguagesAsync(long? tenantId);
 
         /// <summary>
         /// Gets list of all languages available to given tenant (or null for host)
         /// </summary>
         /// <param name="tenantId">TenantId or null for host</param>
-        IReadOnlyList<ApplicationLanguage> GetLanguages(int? tenantId);
+        IReadOnlyList<ApplicationLanguage> GetLanguages(long? tenantId);
 
         /// <summary>
         /// Gets list of all active languages available to given tenant (or null for host)
         /// </summary>
         /// <param name="tenantId">TenantId or null for host</param>
-        IReadOnlyList<ApplicationLanguage> GetActiveLanguages(int? tenantId);
+        IReadOnlyList<ApplicationLanguage> GetActiveLanguages(long? tenantId);
 
         /// <summary>
         /// Adds a new language.
@@ -49,53 +49,53 @@ namespace Abp.Localization
         /// </summary>
         /// <param name="tenantId">Tenant Id or null for host.</param>
         /// <param name="languageName">Name of the language.</param>
-        Task RemoveAsync(int? tenantId, string languageName);
+        Task RemoveAsync(long? tenantId, string languageName);
 
         /// <summary>
         /// Deletes a language.
         /// </summary>
         /// <param name="tenantId">Tenant Id or null for host.</param>
         /// <param name="languageName">Name of the language.</param>
-        void Remove(int? tenantId, string languageName);
+        void Remove(long? tenantId, string languageName);
 
         /// <summary>
         /// Updates a language.
         /// </summary>
         /// <param name="tenantId">Tenant Id or null for host.</param>
         /// <param name="language">The language to be updated</param>
-        Task UpdateAsync(int? tenantId, ApplicationLanguage language);
+        Task UpdateAsync(long? tenantId, ApplicationLanguage language);
 
         /// <summary>
         /// Updates a language.
         /// </summary>
         /// <param name="tenantId">Tenant Id or null for host.</param>
         /// <param name="language">The language to be updated</param>
-        void Update(int? tenantId, ApplicationLanguage language);
+        void Update(long? tenantId, ApplicationLanguage language);
 
         /// <summary>
         /// Gets the default language or null for a tenant or the host.
         /// </summary>
         /// <param name="tenantId">Tenant Id of null for host</param>
-        Task<ApplicationLanguage> GetDefaultLanguageOrNullAsync(int? tenantId);
+        Task<ApplicationLanguage> GetDefaultLanguageOrNullAsync(long? tenantId);
 
         /// <summary>
         /// Gets the default language or null for a tenant or the host.
         /// </summary>
         /// <param name="tenantId">Tenant Id of null for host</param>
-        ApplicationLanguage GetDefaultLanguageOrNull(int? tenantId);
+        ApplicationLanguage GetDefaultLanguageOrNull(long? tenantId);
 
         /// <summary>
         /// Sets the default language for a tenant or the host.
         /// </summary>
         /// <param name="tenantId">Tenant Id of null for host</param>
         /// <param name="languageName">Name of the language.</param>
-        Task SetDefaultLanguageAsync(int? tenantId, string languageName);
+        Task SetDefaultLanguageAsync(long? tenantId, string languageName);
 
         /// <summary>
         /// Sets the default language for a tenant or the host.
         /// </summary>
         /// <param name="tenantId">Tenant Id of null for host</param>
         /// <param name="languageName">Name of the language.</param>
-        void SetDefaultLanguage(int? tenantId, string languageName);
+        void SetDefaultLanguage(long? tenantId, string languageName);
     }
 }
