+++ src/Abp.Zero.Common/MultiTenancy/TenantCacheManagerExtensions.cs
@@ -4,14 +4,14 @@ namespace Abp.MultiTenancy
 {
     public static class TenantCacheManagerExtensions
     {
-        public static ITypedCache<int, TenantCacheItem> GetTenantCache(this ICacheManager cacheManager)
+        public static ITypedCache<long, TenantCacheItem> GetTenantCache(this ICacheManager cacheManager)
         {
-            return cacheManager.GetCache<int, TenantCacheItem>(TenantCacheItem.CacheName);
+            return cacheManager.GetCache<long, TenantCacheItem>(TenantCacheItem.CacheName);
         }
 
-        public static ITypedCache<string, int?> GetTenantByNameCache(this ICacheManager cacheManager)
+        public static ITypedCache<string, long?> GetTenantByNameCache(this ICacheManager cacheManager)
         {
-            return cacheManager.GetCache<string, int?>(TenantCacheItem.ByNameCacheName);
+            return cacheManager.GetCache<string, long?>(TenantCacheItem.ByNameCacheName);
         }
     }
 }
