+++ src/Abp/IUserIdentifier.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// Tenant Id. Can be null for host users.
         /// </summary>
-        int? TenantId { get; }
+        long? TenantId { get; }
 
         /// <summary>
         /// Id of the user.
